Zhaorui Ni's upload firstly for rgb image rendering. Nothing to do with audio for this
Nothing to do with audio for this, please do not change the comment of the TO DO list, otherwise it will lose what I will improve.

If you want to do the audio, please upload yourself and do not modify this. This is nothing to do with audio. Please know what it should do for this java src code.